adding user actions, fix new user edit profile

+++ users/signals.py
@@ -3,22 +3,32 @@ from django.dispatch import receiver
 from django.contrib.auth.models import User
 from .models import Profile
 
-# @receiver(post_save, sender=Profile)
+@receiver(post_save, sender=Profile)
 def create_profile(sender, instance, created, **kwargs):
     if created:
         user = instance
         profile = Profile.objects.create(
-            user=user,
-            username=user.username,
-            email=user.email,
-            name=user.first_name,
+            user = user,
+            username = user.username,
+            email = user.email,
+            name = user.first_name,
         )
 
-# @receiver(post_delete, sender=Profile)
-def deleteuser(sender, instance, **kwargs):
+
+@receiver(post_save, sender=Profile)
+def update_user(sender, instance, created, **kwargs):
+    profile = instance
+    user = profile.user
+    
+    if created == False:
+        user.first_name = profile.name
+        user.username = profile.username
+        user.email = profile.email
+        user.save()
+
+
+@receiver(post_delete, sender=Profile)
+def delete_user(sender, instance, **kwargs):
     user = instance.user
-    user.delete()
-
-
-post_save.connect(create_profile, sender=User)
-post_delete.connect(deleteuser, sender=Profile)
+    if user:
+        user.delete()
